Add ability to hide and unhide menu items
Allow menus to be more dynamic by hiding and unhiding their items. The
meaning of `ConsoleMenu().items` has changed slightly to return a list
of currently unhidden items only so that other classes that use this
value, like `MenuFormatBuilder`, do not break (as long as they are
not trying to modify the list). Also add tests for the new behavior.